requirements.txt
Django~=4.1.7
psycopg2~=2.9.5
python-dotenv~=1.0.0
django-allauth~=0.53.1
weasyprint~=59.0
python-taiga~=1.1.0
celery~=5.3.1
django-celery-beat~=2.5.0
flake8~=6.0.0
gevent~=22.10.2
redis~=4.6.0
channels~=4.0.0
channels-redis~=4.1.0
daphne~=4.0.0